+++ lib/src/smart_dialog.dart
@@ -112,7 +112,7 @@ class SmartDialog {
   ///提供toast示例
   static Future<void> showToast(
     String msg, {
-    Duration time = const Duration(milliseconds: 1500),
+    Duration time = const Duration(milliseconds: 2000),
     alignment: Alignment.bottomCenter,
     //默认消失类型,类似android的toast,toast一个一个展示
     //非默认消失类型,多次点击,后面toast会顶掉前者的toast显示